Time Card: Handle date range in app, not built into filter
The app logic used to assume that the date range was built into the filter so we'd need a "My Time Today" and "My Time This Week", etc. filter to view different time frames.  But we can add date limiting logic to REST query parameters so that the app controls the date range, not the filter.  This reduces the number of filters we need and gives the app more control to do "last month" or "week to date", etc.