Gantt Chart: Treat epic as a parent task
This is harder than it should be because Jira treats epic links specially.  If there were an "epic" link type we could add it to the list of links that get parsed.  But epic link is a custom field and the value is a key (e.g., AAA-1234) not an id.